+++ lib/src/steel_crypt_base.dart
@@ -12,7 +12,7 @@ import 'rsa.dart';
 class SymCrypt {
   core.String type;
   core.String key32;
-  Encrypter encrypter;
+  var encrypter;
 
   SymCrypt (core.String inkey32, [core.String intype = "AES"]) {
     type = intype;
@@ -35,15 +35,16 @@ class SymCrypt {
   } //decrypt base 64 (with iv) and return original
 }
 
+
 class RsaCrypt extends rsa {
   RsaCrypt() : super();
 }
 
+
 class CryptKey {
   var random = Random.secure();
-
   String genKey ([int length = 32]) {
-    var values = List<int>.generate(32, (i) => random.nextInt(256));
+    var values = List<int>.generate(length, (i) => random.nextInt(256));
     return base64Url.encode(values);
   }
 
@@ -52,6 +53,24 @@ class CryptKey {
     return base64Url.encode(values);
   }//gen cryptographically-secure, random key; defaults to length 32
 }
+
+
+class PassCrypt {
+  String hashPass (String salt, String pass, [int length = 32]) {
+    var params = new Pbkdf2Parameters(utf8.encode(salt), 20000, length);
+    var keyDerivator = new KeyDerivator("SHA-512/HMAC/PBKDF2")
+      ..init( params )
+    ;
+    var passBytes = base64.decode(pass);
+    var key = keyDerivator.process( passBytes );
+    return base64.encode(key);
+  }
+  bool checkPassKey (String salt, String plain, String hashed, [int length = 32]) {
+    var hashplain = hashPass(salt, plain, length);
+    return hashplain == hashed;
+  }
+}
+
 
 class HashCrypt {
   core.String type;
@@ -73,13 +92,12 @@ class HashCrypt {
     var val = _tmp.process(bytes);
     return Base64Codec().encode(val);
   }
-  bool checkpass (core.String plain, core.String hashed) {
+  bool checkhash (core.String plain, core.String hashed) {
     var newhash = hash(plain);
     return newhash == hashed;
   }
-  bool checkpassHMAC (core.String plain, core.String hashed, core.String key) {
+  bool checkhashHMAC (core.String plain, core.String hashed, core.String key) {
     var newhash = hashHMAC(plain, key);
     return newhash == hashed;
   }
 }
-